Fix capitalization error in setUp and tearDown
